feat(transactions): rework how transactions work
BREAKING CHANGE: upgrade to lodash@4 changing many methods that are used internally. Splits payment transactions into multiple groups.
Agreements are created with setup_fee and recurring payments starting in 'frequency' interval. That is a month or a year from a start
date. Executing agreement now automatically cancels any existing paid agreement. It must be noted, that we can't change an agreement,
while it's current status is said to be pending, because, otherwise, cancelling will fail.

* Adds sale.get, transaction.get, agreement.get endpoints
* Adds transaction.common, which lists all types of transactions and can be filtered in many ways
* Rework sale.execute and how models are added in order to avoid duplicate additions of models
* Add constants instead of directly using strings for keys in redis
* Move mapValues to serialize/deserialize utils for redis
* Add periodic syncer of sales and agreement transactions
* On startup, the service will pull possibly missing transactions and update agreements