refactor: remove any plan mutatating queries on paypal side
Reason is that Paypal is a piece of shit. It doesn't support mutating plan template,
which has nothing to do with the actual agreement that is created later.
To update a plan - one needs to rename alias to something else, turn it to hidden, then create
new plan with appropriate values